+++ les10/app/helpers/classes.js
@@ -1,0 +1,4 @@
+const classes = ['Warrior', 'Mage', 'Warlock', 'Knight', 'Druid', 'Monk', 'Prophet', 'Hunter',
+    'Rogue', 'Paladin', 'Shaman', 'Priest', 'Death Knight', 'Demon Hunter'];
+
+export default classes;
