+++ les10/app/helpers/cookie.js
@@ -1,0 +1,18 @@
+export default class Cookie{
+    static set(cname, cvalue, exdays) {
+        const d = new Date();
+        d.setTime(d.getTime() + (exdays * 24 * 60 * 60 * 1000));
+
+        const expires = `expires=${d.toUTCString()}`;
+        document.cookie = `${cname}=${cvalue};${expires};path=/`;
+    }
+    static get(cname) {
+        const name = `${cname}=`;
+        const coo = document.cookie.split(';').find(el => (el.trim()).startsWith(name));
+
+        return coo ? coo.trim().slice(name.length) : '';
+    }
+    static del(cname){
+        document.cookie = `${cname}=; expires=Thu, 01 Jan 1970 00:00:00 UTC; path=/;`;
+    }
+}
