+++ les10/app/helpers/races.js
@@ -1,0 +1,60 @@
+const races = [
+    {
+        name : 'Human',
+        url : 'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSRoOJ3wiMGoC1y2NQPkKRCKFcyMD1qrK0n7g&usqp=CAU'
+    },
+    {
+        name : 'Elven',
+        url : 'https://i.etsystatic.com/9879972/r/il/494959/1531730199/il_570xN.1531730199_8ero.jpg'
+    },
+    {
+        name : 'Orc',
+        url : 'https://cdn.dribbble.com/users/2043038/screenshots/4708721/orc.png'
+    },
+    {
+        name : 'Dragon',
+        url : 'https://media.istockphoto.com/vectors/dragon-icon-with-fire-symbol-vector-id1191986285?b=1&k=6&m=1191986285&s=612x612&w=0&h=XGvPRd8BjmkRgv4eQUFq4tHNoMYReL9tm8qEf7nUoxA='
+    },
+    {
+        name : 'Dwarf',
+        url : 'https://www.creativefabrica.com/wp-content/uploads/2019/04/Garden-gnome-line-art-drawing-4-580x386.jpg'
+    },
+    {
+        name : 'Undead',
+        url : 'https://i.pinimg.com/originals/74/e2/21/74e22176c7c930ce2518bb2e45cebc31.jpg'
+    },
+    {
+        name : 'Troll',
+        url : 'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRdGHuvLLMoqvCKo0XzHVRoxlxR_205AkAOSA&usqp=CAU'
+    },
+    {
+        name : 'Night Elf',
+        url : 'https://i.pinimg.com/originals/8b/eb/97/8beb977853c4819b27c1e33c3e45c7ca.jpg'
+    },
+    {
+        name : 'Draenei',
+        url : 'https://lh3.googleusercontent.com/proxy/tmXkMMuYCH3zgsfhp8wQxX3ifk0fFu1xr8Eoa0Tfp_BdNq8OJJ_kCapgMZBwU482pSb9a8XjxJvLv9NFfxMlGl7Zxv8_Mdg664ITCiDtYhxy1eV2XiUzu_4pLFR0Te9x6RtV1pP3ARvJcTrKNh6IAKUzvcPotYWR3JPZfSM'
+    },
+    {
+        name : 'Worgen',
+        url : 'https://disciplinaryaction.files.wordpress.com/2012/04/icon_wooden_worgen.png'
+    },
+    {
+        name : 'Pandaren',
+        url : 'https://mir-s3-cdn-cf.behance.net/projects/404/e147ee91583457.Y3JvcCw5MDcsNzEwLDYwLDA.png'
+    },
+    {
+        name : 'Tauren',
+        url : 'https://www.kindpng.com/picc/m/28-289775_wow-tauren-symbol-hd-png-download.png'
+    },
+    {
+        name : 'Blood Elf',
+        url : 'https://i.pinimg.com/originals/7c/37/bf/7c37bf2f28aefca6e5335da6d1a933fe.jpg'
+    },
+    {
+        name : 'Goblin',
+        url : 'https://icon-library.com/images/goblin-icon/goblin-icon-6.jpg'
+    }
+];
+
+export default races;
